Address workspace lifecycle review feedback
